+++ frontend/wailsjs/go/main/App.d.ts
@@ -77,6 +77,8 @@ export function GetActivities():Promise<Record<string, string>>;
 
 export function GetAgents():Promise<Array<main.AgentInfo>>;
 
+export function GetAppLog():Promise<main.AppLog>;
+
 export function GetBackgroundAgentLogs(arg1:string):Promise<string>;
 
 export function GetBackups():Promise<Array<main.BackupInfo>>;
@@ -194,6 +196,8 @@ export function MoveGroup(arg1:string,arg2:number):Promise<void>;
 export function MoveSessionToIndex(arg1:string,arg2:number):Promise<void>;
 
 export function MoveTask(arg1:string,arg2:string,arg3:string):Promise<void>;
+
+export function OpenAppLogFolder():Promise<void>;
 
 export function OpenFolder(arg1:string):Promise<void>;